dom/html/test/file_iframe_sandbox_j_if3.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Tests for Bug 766282</title>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

</head>
<script type="text/javascript">
  function ok(result, desc) {
    window.parent.ok_wrapper(result, desc);
  }

  function doStuff() {
    // Try to open a new window via showModalDialog().
    // The window we try to open closes itself once it opens.
    try {
      window.showModalDialog("file_iframe_sandbox_open_window_pass.html");
    } catch(e) {
      ok(false, "iframes sandboxed with allow-popups and allow-modals should be able to open a modal dialog");
    }
  }
</script>
<body onLoad="doStuff()">
  I am sandboxed but with "allow-popups allow-scripts allow-same-origin"
</body>
</html>
